% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{intensityDataset}
\alias{intensityDataset}
\title{Plot the intensity of values}
\usage{
intensityDataset(dataset, rankList, position, alpha)
}
\arguments{
\item{dataset}{Dataset containing numeric values}

\item{rankList}{List of ranked motifs}

\item{position}{Select by an integer a motif with his position}

\item{alpha}{Number of letter used to do the encode}
}
\value{
Pixelated dataset
}
\description{
Display the intensity of values and higthlight one motif
}
\examples{
#Launch all the workflow
#Plot the result
D  <- STMotif::example_dataset
DS <- NormSAX(STMotif::example_dataset,7)
stmotifs <- SearchSTMotifs(D,DS,3,7,10,10,3,7)
rstmotifs <- RankSTMotifs(stmotifs)
intensityDataset(dataset = STMotif::example_dataset, rankList = rstmotifs, 1, 7)
}
