% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subFunction.R
\name{plot.series}
\alias{plot.series}
\title{Plot series lines with facet grid for spatial-time series}
\usage{
\method{plot}{series}(series, label_series = "", label_x = "", label_y = "")
}
\arguments{
\item{series}{Melted data frame with x, value, color, and variable columns.}

\item{label_series}{Optional label for series.}

\item{label_x}{Optional x-axis label.}

\item{label_y}{Optional y-axis label.}
}
\description{
Plot series lines with facet grid for spatial-time series
}
