% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{runVisualization}
\alias{runVisualization}
\title{Interactive visualization}
\usage{
runVisualization(dataset, rankList, alpha)
}
\arguments{
\item{dataset}{Dataset containing numeric values}

\item{rankList}{List of ranked motifs}

\item{alpha}{Number of letters used to do the encode}
}
\description{
Launch a process to have an interactive visualization
}
\examples{
\dontrun{
#Launch all the workflow
D  <- STMotif::example_dataset
DS <- NormSAX(STMotif::example_dataset,7)
stmotifs <- SearchSTMotifs(D,DS,3,7,10,10,3,7)
rstmotifs <- RankSTMotifs(stmotifs)
#Launch the process
runVisualization(dataset = STMotif::example_dataset, rstmotifs, 7)
}
}
