% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STMotif.R
\docType{package}
\name{STMotif}
\alias{STMotif}
\alias{STMotif-package}
\title{Package STMotif}
\description{
This package `STSMotifs` allows to perform a research of motif into spatial time series. The main purpose is to find a way to handle the issue of large amouts of data. The package offers a way to do this research quickly and efficiently.
}
\details{
To have more information about the package : \href{../doc/motifs-discovery.html}{PACKAGE STMOTIF}
}
