% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunction.R
\name{STSADatasetAdjust}
\alias{STSADatasetAdjust}
\title{Adjust a Dataset}
\usage{
STSADatasetAdjust(D, tb, sb)
}
\arguments{
\item{D}{Dataset containing numeric values.}

\item{tb}{Temporal block size (number of rows per block).}

\item{sb}{Spatial block size (number of columns per block).}
}
\value{
Dataset with rows and columns trimmed to be divisible by
  \code{tb} and \code{sb}, respectively.
}
\description{
Adjusts the dimensions of a dataset so that it can be evenly divided
into spatio-temporal blocks of size \code{tb} x \code{sb}.
}
\examples{
D <- STSADatasetAdjust(STMotif::example_dataset, 20, 12)
}
