% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunction.R
\name{STSRankTightSTMotifs}
\alias{STSRankTightSTMotifs}
\title{Rank the motifs}
\usage{
STSRankTightSTMotifs(sttightmotifs)
}
\arguments{
\item{sttightmotifs}{List of motifs after removing isolated motifs}
}
\value{
The ranked version of the list provided as input.
}
\description{
Rank motifs by their quality
}
\examples{
#Generation of candidates
candidates <- STSIdentifyCandidateSTMotifs(STMotif::example_dataset, 10, 10, 7, 3)

#Filter the selected motifs by their occurences
stmotifs <- STSIdentifySTMotifs(candidates, 1, 1)

#Remove the isolated motifs
sttightmotifs <- STSIdentifyTightSTMotifs(stmotifs, candidates$rectangles)

#Order the result list
rank <- STSRankTightSTMotifs(sttightmotifs)
}
