% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_dataset.R
\docType{data}
\name{example_dataset}
\alias{example_dataset}
\title{Example of dataset}
\format{An object of class \code{data.frame} with 100 rows and 50 columns.}
\usage{
example_dataset
}
\description{
Toy example to launch functions.
The dimensions of the dataset are 100 rows and 50 columns and this dataset contains 50 spatial-time series.
}
\keyword{example_dataset}
