% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_dataset.R
\docType{data}
\name{example_dataset}
\alias{example_dataset}
\title{Example of dataset}
\format{The dimensions of the dataset are 20 rows and 12 columns and this dataset contains 12 spatial-time series.}
\usage{
example_dataset
}
\description{
Toy example to launch functions.
}
\keyword{example_dataset}
