% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunction.R
\name{NormSAX}
\alias{NormSAX}
\title{Normalize the data and SAX indexing}
\usage{
NormSAX(D, a)
}
\arguments{
\item{D}{Dataset containing numeric values}

\item{a}{Number of letters use to encode}
}
\value{
A normalized and encoded dataset for a given alphabet a
}
\description{
Normalize the data and SAX indexing
}
\examples{
#Normalization and Sax Dataset
DS <- NormSAX(STMotif::example_dataset, 5)
}
