% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunction.R
\name{RankSTMotifs}
\alias{RankSTMotifs}
\title{Rank the STmotifs
Rank motifs by their quality}
\usage{
RankSTMotifs(stmotifs)
}
\arguments{
\item{stmotifs}{List of identified motifs}
}
\value{
The ranked version of the identified list of motifs
}
\description{
Rank the STmotifs
Rank motifs by their quality
}
\examples{
#Search for Spatial-time Motifs
D  <- STMotif::example_dataset
DS <- NormSAX(STMotif::example_dataset,7)
stmotifs <- SearchSTMotifs(D,DS,3,7,10,10,3,7)
rstmotifs <- RankSTMotifs(stmotifs)
}
