% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{runVisualization}
\alias{runVisualization}
\title{Interactive visualization}
\usage{
runVisualization(dataset, rankList, alpha)
}
\arguments{
\item{dataset}{Dataset containing numeric values}

\item{rankList}{List of ranked motifs}

\item{alpha}{Number of letter used to do the encode}
}
\description{
Launch a process to have an interactive visualization
}
\examples{
\dontrun{
#Launch all the workflow
candidates <- STSIdentifyCandidateSTMotifs(STMotif::example_dataset, 10, 10, 7, 3)
stmotifs <- STSIdentifySTMotifs(candidates, 1, 1)
sttightmotifs <- STSIdentifyTightSTMotifs(stmotifs, candidates$rectangles)
rank <- STSRankTightSTMotifs(sttightmotifs)

#Launch the process
runVisualization(dataset = STMotif::example_dataset, rank, 7)
}
}
