% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STMotif.R
\docType{package}
\name{STMotif}
\alias{STMotif}
\alias{STMotif-package}
\title{Package STMotif}
\description{
This package `STSMotifs` allows to identify motifs in spatial-time series. A motif is a previously unknown subsequence of a (spatial) time series with relevant number of occurrences. For this purpose, the Combined Series Approach (CSA) is used.
}
\details{
To have more information about the package : \href{../doc/STMotif.html}{PACKAGE STMOTIF}
}
