% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunction.R
\name{STSADatasetAdjust}
\alias{STSADatasetAdjust}
\title{Adjust a Dataset
Adjust the dimensions of a dataset to build the blocks}
\usage{
STSADatasetAdjust(D, tb, sb)
}
\arguments{
\item{D}{Dataset containing numeric values}

\item{tb}{Temporal block size}

\item{sb}{Spatial block size}
}
\value{
Dataset adjusted to build the blocks.
}
\description{
Adjust a Dataset
Adjust the dimensions of a dataset to build the blocks
}
\examples{
#Adjust a block
D <- STSADatasetAdjust(STMotif::example_dataset, 20, 12)
}
