% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{top5motifs}
\alias{top5motifs}
\title{Plot the 5 motifs}
\usage{
top5motifs(dataset, rankList, alpha)
}
\arguments{
\item{dataset}{Dataset containing numeric values}

\item{rankList}{List of ranked motifs}

\item{alpha}{Number of letters used to do the encode}
}
\value{
Pixelated dataset
}
\description{
Display the intensity of values and higthlight the top five motifs
}
\examples{
#Launch all the workflow
D  <- STMotif::example_dataset
DS <- NormSAX(STMotif::example_dataset,7)
stmotifs <- SearchSTMotifs(D,DS,3,7,10,10,3,7)
rstmotifs <- RankSTMotifs(stmotifs)
#Plot the result
top5motifs(dataset = STMotif::example_dataset, rankList = rstmotifs, alpha = 7)
}
